+++ src/routes/global/health_check.ts
@@ -1,0 +1,14 @@
+import { URL, RequestResponse } from '../../types/url';
+
+export interface HealthCheck {
+    '/ping': URL<{
+        // GET request
+        get: RequestResponse<{
+            status: '200';
+            body: {
+                status: 'success';
+                message: string;
+            };
+        }>;
+    }>;
+}
