+++ src/routes/global/version.ts
@@ -1,0 +1,14 @@
+import { URL, RequestResponse } from '../../types/url';
+
+export interface Version {
+    '/version': URL<{
+        // GET request
+        get: RequestResponse<{
+            status: '200';
+            body: {
+                status: 'success';
+                message: string;
+            };
+        }>;
+    }>;
+}
